@@ -1,6 +1,6 @@
 {
   "name": "nergal.github.com",
-  "version": "1.0.0",
+  "version": "0.0.2",
   "description": "",
   "main": "index.html",
   "scripts": {
@@ -28,27 +28,27 @@
     }
   },
   "devDependencies": {
-    "babel-core": "^6.18.2",
-    "babel-eslint": "6",
-    "babel-loader": "^6.2.8",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
+    "babel-core": "6.24.1",
+    "babel-eslint": "8.0.1",
+    "babel-loader": "7.1.2",
+    "babel-preset-env": "1.6.1",
+    "babel-preset-react": "6.24.1",
     "bootstrap": "3",
-    "css-loader": "^0.26.0",
-    "eslint": "3.x",
-    "eslint-loader": "^1.6.1",
-    "file-loader": "^0.9.0",
-    "json-loader": "^0.5.4",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "react": "15.4.0",
-    "react-bootstrap": "^0.30.7",
-    "react-dom": "15.4.0",
-    "react-ga": "^2.1.2",
-    "react-string-replace": "^0.4.0",
-    "style-loader": "^0.13.1",
-    "svg-loader": "^0.0.2",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.3"
+    "css-loader": "0.28.7",
+    "eslint": "4.9.0",
+    "eslint-loader": "1.9.0",
+    "file-loader": "1.1.5",
+    "json-loader": "0.5.7",
+    "less": "2.7.2",
+    "less-loader": "4.0.5",
+    "prop-types": "15.6.0",
+    "react": "16.0.0",
+    "react-bootstrap": "0.31.3",
+    "react-dom": "16.0.0",
+    "react-ga": "2.3.5",
+    "react-string-replace": "0.4.1",
+    "style-loader": "0.19.0",
+    "url-loader": "0.6.2",
+    "webpack": "3.8.1"
   }
 }
