@@ -1,6 +1,7 @@
 import '../assets/styles.less';
 
-import React, {Component, PropTypes} from 'react';
+import React, {Component} from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import ReactGA from 'react-ga';
 
