@@ -1,4 +1,5 @@
-import {Component, PropTypes} from 'react';
+import {Component} from 'react';
+import PropTypes from 'prop-types';
 import {renderGeneral, renderSidebar, renderOverview, renderAdditionally, renderExperience} from './sections';
 
 export default class SectionFactory extends Component {
